requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
asgiref==3.9.1
    # via django
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
django==5.2.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
openai==1.102.0
    # via -r requirements.in
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-telegram-bot==22.3
    # via -r requirements.in
regex==2025.7.34
    # via tiktoken
requests==2.32.5
    # via tiktoken
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.3
    # via django
tiktoken==0.11.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via django
urllib3==2.5.0
    # via requests
